+++ src/board/BoardToolbar.tsx
@@ -1,0 +1,76 @@
+import {
+  ArrowDownToolbarItem,
+  ArrowLeftToolbarItem,
+  ArrowRightToolbarItem,
+  ArrowToolbarItem,
+  ArrowUpToolbarItem,
+  CheckBoxToolbarItem,
+  CloudToolbarItem,
+  DefaultToolbar,
+  DiamondToolbarItem,
+  DrawToolbarItem,
+  EllipseToolbarItem,
+  EraserToolbarItem,
+  FrameToolbarItem,
+  HandToolbarItem,
+  HeartToolbarItem,
+  HexagonToolbarItem,
+  HighlightToolbarItem,
+  LaserToolbarItem,
+  LineToolbarItem,
+  OvalToolbarItem,
+  RectangleToolbarItem,
+  RhombusToolbarItem,
+  SelectToolbarItem,
+  StarToolbarItem,
+  TextToolbarItem,
+  TldrawUiMenuItem,
+  TriangleToolbarItem,
+  XBoxToolbarItem,
+} from 'tldraw'
+import { useBoardActions } from '../context/BoardActionsContext'
+
+export function BoardToolbar() {
+  const { onAddNote, canAddNote } = useBoardActions()
+
+  return (
+    <DefaultToolbar>
+      <SelectToolbarItem />
+      <HandToolbarItem />
+      <DrawToolbarItem />
+      <EraserToolbarItem />
+      <TldrawUiMenuItem
+        id="add-tip-note"
+        label="Add note"
+        icon="tool-note"
+        kbd="n"
+        onSelect={() => {
+          if (canAddNote) onAddNote()
+        }}
+        disabled={!canAddNote}
+      />
+      <ArrowToolbarItem />
+      <TextToolbarItem />
+      <RectangleToolbarItem />
+      <EllipseToolbarItem />
+      <TriangleToolbarItem />
+      <DiamondToolbarItem />
+      <HexagonToolbarItem />
+      <OvalToolbarItem />
+      <RhombusToolbarItem />
+      <StarToolbarItem />
+      <CloudToolbarItem />
+      <HeartToolbarItem />
+      <XBoxToolbarItem />
+      <CheckBoxToolbarItem />
+      <ArrowLeftToolbarItem />
+      <ArrowUpToolbarItem />
+      <ArrowDownToolbarItem />
+      <ArrowRightToolbarItem />
+      <LineToolbarItem />
+      <HighlightToolbarItem />
+      <LaserToolbarItem />
+      <FrameToolbarItem />
+    </DefaultToolbar>
+  )
+}